move publish action and bulk action
