+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/entity-actions/manifests.ts
@@ -2,7 +2,6 @@ import { UMB_DOCUMENT_DETAIL_REPOSITORY_ALIAS, UMB_DOCUMENT_ITEM_REPOSITORY_ALIA
 import { UMB_DOCUMENT_ENTITY_TYPE } from '../entity.js';
 import {
 	UMB_USER_PERMISSION_DOCUMENT_DELETE,
-	UMB_USER_PERMISSION_DOCUMENT_PUBLISH,
 	UMB_USER_PERMISSION_DOCUMENT_UNPUBLISH,
 } from '../user-permissions/constants.js';
 import { manifests as createBlueprintManifests } from './create-blueprint/manifests.js';
@@ -37,29 +36,6 @@ const entityActions: Array<UmbExtensionManifest> = [
 			},
 			{
 				alias: UMB_ENTITY_IS_TRASHED_CONDITION_ALIAS,
-			},
-		],
-	},
-	{
-		type: 'entityAction',
-		kind: 'default',
-		alias: 'Umb.EntityAction.Document.Publish',
-		name: 'Publish Document Entity Action',
-		weight: 600,
-		api: () => import('./publish.action.js'),
-		forEntityTypes: [UMB_DOCUMENT_ENTITY_TYPE],
-		meta: {
-			icon: 'icon-globe',
-			label: '#actions_publish',
-			additionalOptions: true,
-		},
-		conditions: [
-			{
-				alias: 'Umb.Condition.UserPermission.Document',
-				allOf: [UMB_USER_PERMISSION_DOCUMENT_PUBLISH],
-			},
-			{
-				alias: UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
 			},
 		],
 	},
