+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/entity-bulk-actions/manifests.ts
@@ -14,29 +14,6 @@ export const entityBulkActions: Array<ManifestEntityBulkAction> = [
 	{
 		type: 'entityBulkAction',
 		kind: 'default',
-		alias: 'Umb.EntityBulkAction.Document.Publish',
-		name: 'Publish Document Entity Bulk Action',
-		weight: 50,
-		api: () => import('./publish/publish.action.js'),
-		meta: {
-			icon: 'icon-globe',
-			label: '#actions_publish',
-		},
-		forEntityTypes: [UMB_DOCUMENT_ENTITY_TYPE],
-		conditions: [
-			{
-				alias: UMB_COLLECTION_ALIAS_CONDITION,
-				match: UMB_DOCUMENT_COLLECTION_ALIAS,
-			},
-			{
-				alias: 'Umb.Condition.UserPermission.Document',
-				allOf: [UMB_USER_PERMISSION_DOCUMENT_PUBLISH],
-			},
-		],
-	},
-	{
-		type: 'entityBulkAction',
-		kind: 'default',
 		alias: 'Umb.EntityBulkAction.Document.Unpublish',
 		name: 'Unpublish Document Entity Bulk Action',
 		weight: 40,
