+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/publish/entity-action/index.ts
@@ -1,0 +1,1 @@
+export * from './publish.action.js';
