+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/publish/entity-action/publish.action.ts
@@ -1,7 +1,5 @@
-import { UmbDocumentDetailRepository } from '../repository/index.js';
-import { UmbDocumentPublishingRepository } from '../publishing/index.js';
-import { UMB_DOCUMENT_PUBLISH_MODAL } from '../publishing/publish/modal/constants.js';
-import type { UmbDocumentVariantOptionModel } from '../types.js';
+import { type UmbDocumentVariantOptionModel } from '../../../types.js';
+import { UMB_DOCUMENT_PUBLISH_MODAL } from '../modal/constants.js';
 import { UMB_APP_LANGUAGE_CONTEXT, UmbLanguageCollectionRepository } from '@umbraco-cms/backoffice/language';
 import type { UmbEntityActionArgs } from '@umbraco-cms/backoffice/entity-action';
 import { UmbEntityActionBase, UmbRequestReloadStructureForEntityEvent } from '@umbraco-cms/backoffice/entity-action';
@@ -10,6 +8,8 @@ import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UMB_MODAL_MANAGER_CONTEXT } from '@umbraco-cms/backoffice/modal';
 import { UMB_ACTION_EVENT_CONTEXT } from '@umbraco-cms/backoffice/action';
 import { UMB_CURRENT_USER_CONTEXT } from '@umbraco-cms/backoffice/current-user';
+import { UmbDocumentDetailRepository } from '../../../repository/index.js';
+import { UmbDocumentPublishingRepository } from '../../repository/index.js';
 
 export class UmbPublishDocumentEntityAction extends UmbEntityActionBase<never> {
 	constructor(host: UmbControllerHost, args: UmbEntityActionArgs<never>) {
