+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/publish/manifests.ts
@@ -1,3 +1,9 @@
+import { manifests as entityActionManifests } from './entity-action/manifests.js';
+import { manifests as entityBulkActionManifests } from './entity-bulk-action/manifests.js';
 import { manifests as modalManifests } from './modal/manifests.js';
 
-export const manifests: Array<UmbExtensionManifest> = [...modalManifests];
+export const manifests: Array<UmbExtensionManifest> = [
+	...entityActionManifests,
+	...entityBulkActionManifests,
+	...modalManifests,
+];
